+++ CHANGELOG.md
@@ -385,6 +385,7 @@
   - added: session_id & session_url -> session.json for reconnecting to existing browser sessions
   **2.18.1**
   - debugged: redundant category asking
+  - debugging: local
 
 ----------------------------------------
 
@@ -393,8 +394,6 @@
   - adding tab requires a file lock between onlysnarf sessions to maintain the correct active tab
   - debugged remote: debug refresh to stay open -> test javascript refresh injection
 
-
-  - need to debug local files
   
 
   **x.x.x**